runtime: Move module/test.rs to it's own file
